refactor modules layout
