+++ .gitmodules
@@ -1,0 +1,9 @@
+[submodule "modules/ap98j3envoubi3fco1kc"]
+	path = modules/ap98j3envoubi3fco1kc
+	url = git@github.com:exorde-labs/ap98j3envoubi3fco1kc.git
+[submodule "modules/ch4875eda56be56000ac"]
+	path = modules/ch4875eda56be56000ac
+	url = git@github.com:exorde-labs/ch4875eda56be56000ac.git
+[submodule "modules/a7df32de3a60dfdb7a0b"]
+	path = modules/a7df32de3a60dfdb7a0b
+	url = git@github.com:exorde-labs/a7df32de3a60dfdb7a0b.git
