--- data/exorde_data/scraping.py
@@ -1,7 +1,0 @@
-scraping_modules = {
-    "reddit": "ap98j3envoubi3fco1kc",
-    "twitter": "a7df32de3a60dfdb7a0b",
-    "4chan": "ch4875eda56be56000ac",
-}
-
-__all__ = ["scraping_modules"]
