--- data/schema.json
@@ -1,55 +1,0 @@
-{
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "https://raw.githubusercontent.com/exorde-labs/exorde-client/main/data/schema.json",
-    "type": "object",
-    "properties": {
-        "created_at": {
-            "description": "ISO8601/RFC3339 Date of creation of the item",
-            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(.[0-9]{1,6})?Z$",
-            "type": "string"
-        },
-        "title": {
-            "description": "Headline of the item",
-            "type": "string"
-        },
-        "content": {
-            "description": "Text body of the item",
-            "type": "string"
-        },
-        "summary": {
-            "description": "Short version of the content",
-            "type": "string"
-        },
-        "picture": {
-            "description": "Uniform-Resource-Locator",
-            "pattern": "^https?:\\/\\/(?:www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,32}\\b(?:[-a-zA-Z0-9()@:%_\\+.~#?&\\/=]*)$",
-            "type": "string"
-        },
-        "author": {
-            "description": "SHA1 encoding of the username assigned as creator of the item on its source platform",
-            "type": "string"
-        },
-        "external_id": {
-            "description": "Identifier used by source",
-            "type": "string"
-        },
-        "external_parent_id": {
-            "description": "Identifier of parent item, as used by source",
-            "type": "string"
-        },
-        "domain": {
-            "description": "Domain name on which the item was retrieved",
-            "type": "string"
-        },
-        "url": {
-            "description": "Uniform-Resource-Locator",
-            "pattern": "^https?:\\/\\/(?:www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,32}\\b(?:[-a-zA-Z0-9()@:%_\\+.~#?&\\/=]*)$",
-            "type": "string"
-        }
-    },
-    "required": [
-        "created_at",
-        "domain",
-        "url"
-    ]
-}
