--- data/scraping/ap98j3envoubi3fco1kc/README.md
@@ -1,12 +1,0 @@
-<img src="https://img.shields.io/badge/exorde-data-blue?style=for-the-badge" />
-
-#### This package is part of [exorde_data]().
-
-- :warning: Exorde data is a modular scraping module that propose to it's users to install a wide range of sub-modules.
-
-- ⚠️  To avoid possible security risk with name pollution, each of our sub-modules is released with a random uuid[:15] name.
-
-**Instalation should be done trough the exorde_data module**
-```bash
-pip install exorde_data[reddit]
-```
