--- data/scraping/ap98j3envoubi3fco1kc/setup.py
@@ -1,9 +1,0 @@
-from setuptools import find_packages, setup
-
-setup(
-    name="ap98j3envoubi3fco1kc",
-    version="0.0.3",
-    packages=find_packages(),
-    install_requires=["lxml", "exorde_data", "aiohttp"],
-    extras_require={"dev": ["pytest", "pytest-cov", "pytest-asyncio"]},
-)
